requirements.txt
Django
django-allauth
flake8
mysqlclient
alembic
grpcio
grpcio-tools
sqlalchemy
